pyproject.toml
[tool.poetry]
name = "pochildren-story-generator"
version = "0.1.0"
description = "Generator for fairy tales"
authors = ["Maksym Drahun <maksym.drahun@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "3.0.3"
langchain-core = "0.2.35"
langchain-google-genai = "1.0.10"
google-api-python-client = "^2.70.0"
requests = "^2.32.3"
python-dotenv = "^1.0.0"
google-generativeai = "0.7.2"
gunicorn = "23.0.0"